Fixed admin delete server error by handling exception

+++ wosoapi/accounts/admin_task.py
@@ -7,4 +7,8 @@ User = get_user_model()
 
 def delete_unverified_users(days=7):
     cutoff = timezone.now() - timedelta(days=days)
-    User.objects.filter(is_active=False, date_joined__lt=cutoff).delete()
+    if User.objects.filter(is_active=False, date_joined__lt=cutoff).exists():
+        User.objects.filter(is_active=False, date_joined__lt=cutoff).delete()
+        return "Unverified users deleted successfully."
+    else:
+        return "No unverified users to delete."
